MeterPool

+++ Aban360.MeterPool.Domain/Features/Management/Dtos/Commands/ReadingPeriodCreateDto.cs
@@ -4,6 +4,6 @@
     {
         public string Title { get; set; } = null!;
         public short ReadingPeriodTypeId { get; set; }
-        public int ClientOrder { get; set; }
+        public short ClientOrder { get; set; }
     }
 }
