+++ Aban360.MeterPool.Domain/Features/Management/Dtos/Commands/ReadingPeriodTypeUdpateDto.cs
@@ -5,7 +5,7 @@
         public short Id { get; set; }
         public string Title { get; set; } = null!;
         public short Days { get; set; }
-        public int ClientOrder { get; set; }
+        public short ClientOrder { get; set; }
         public bool IsEnabled { get; set; }
         public short HeadquartersId { get; set; }
     }
