+++ Aban360.MeterPool.Domain/Features/Management/Dtos/Commands/ReadingPeriodUdpateDto.cs
@@ -5,6 +5,6 @@
         public short Id { get; set; }
         public string Title { get; set; } = null!;
         public short ReadingPeriodTypeId { get; set; }
-        public int ClientOrder { get; set; }
+        public short ClientOrder { get; set; }
     }
 }
