+++ Aban360.MeterPool.Domain/Features/Management/Entities/ReadingPeriod.cs
@@ -8,6 +8,6 @@ public class ReadingPeriod
     public short Id { get; set; }
     public string Title { get; set; } = null!;
     public short ReadingPeriodTypeId { get; set; }
-    public int ClientOrder { get; set; }
+    public short ClientOrder { get; set; }
     public virtual ReadingPeriodType ReadingPeriodType { get; set; } = null!;
 }
